index-node: Pass auth as Bearer token
