+++ server/index-node/src/resolver.rs
@@ -3,7 +3,6 @@ use std::collections::BTreeMap;
 use std::convert::TryInto;
 use web3::types::{Address, H256};
 
-use crate::POI_PROTECTION;
 use graph::blockchain::{Blockchain, BlockchainKind};
 use graph::data::subgraph::features::detect_features;
 use graph::data::subgraph::{status, MAX_SPEC_VERSION};
@@ -15,11 +14,14 @@ use graph::{
 };
 use graph_graphql::prelude::{a, ExecutionContext, Resolver};
 
+use crate::auth::POI_PROTECTION;
+
 /// Resolver for the index node GraphQL API.
 pub struct IndexNodeResolver<S, R> {
     logger: Logger,
     store: Arc<S>,
     link_resolver: Arc<R>,
+    bearer_token: Option<String>,
 }
 
 impl<S, R> IndexNodeResolver<S, R>
@@ -27,12 +29,18 @@ where
     S: Store,
     R: LinkResolver,
 {
-    pub fn new(logger: &Logger, store: Arc<S>, link_resolver: Arc<R>) -> Self {
+    pub fn new(
+        logger: &Logger,
+        store: Arc<S>,
+        link_resolver: Arc<R>,
+        bearer_token: Option<String>,
+    ) -> Self {
         let logger = logger.new(o!("component" => "IndexNodeResolver"));
         Self {
             logger,
             store,
             link_resolver,
+            bearer_token,
         }
     }
 
@@ -170,11 +178,8 @@ where
         let mut indexer = field
             .get_optional::<Address>("indexer")
             .expect("Invalid indexer");
-        let access_token = field
-            .get_optional::<String>("accessToken")
-            .expect("Invalid accessToken");
 
-        if !POI_PROTECTION.validate_access_token(access_token.as_deref()) {
+        if !POI_PROTECTION.validate_access_token(self.bearer_token.as_deref()) {
             // Let's sign the POI with a zero'd address when the access token is
             // invalid.
             indexer = Some(Address::zero());
@@ -500,6 +505,7 @@ where
             logger: self.logger.clone(),
             store: self.store.clone(),
             link_resolver: self.link_resolver.clone(),
+            bearer_token: self.bearer_token.clone(),
         }
     }
 }
