+++ server/index-node/src/schema.graphql
@@ -18,7 +18,6 @@ type Query {
     blockNumber: Int!
     blockHash: Bytes!
     indexer: Bytes
-    accessToken: String
   ): Bytes
   subgraphFeatures(subgraphId: String!): SubgraphFeatures!
   entityChangesInBlock(subgraphId: String!, blockNumber: Int!): EntityChanges!
